KIET_AID_Team_2/Backend/requirements.txt
# ===================== BACKEND REQUIREMENTS =====================

# -------- Core FastAPI --------
fastapi==0.104.1
# Optional: Async email backend for emergency notifications
fastapi-mail
uvicorn==0.24.0
python-multipart==0.0.6
pydantic[email]==2.5.0

# -------- Machine Learning / NLP --------
torch==2.1.0
transformers==4.35.0
peft==0.7.1
accelerate==0.24.1
sentencepiece==0.1.99

# -------- Audio / Voice Processing --------
SpeechRecognition==3.10.0
pydub==0.25.1
librosa==0.10.0

# -------- Scientific & Utilities --------
numpy==1.24.3
scipy==1.11.4
python-dotenv==1.0.0

# -------- Database --------
pymongo==4.6.1
dnspython==2.4.2

# -------- Production Server --------
gunicorn==21.2.0

# optional dependency that enables WebSocket support for uvicorn
# you can also install uvicorn with the "[standard]" extras which pulls this
websockets>=11.0.3

# -------- HTTP client used by Gemini integration --------
httpx==0.29.4
 